# perf(pycodestyle): Introduce TokenKind