--- a/crates/ruff/src/rules/pycodestyle/rules/logical_lines/indentation.rs
+++ b/crates/ruff/src/rules/pycodestyle/rules/logical_lines/indentation.rs
@@ -1,8 +1,8 @@
 use ruff_diagnostics::DiagnosticKind;
 use ruff_diagnostics::Violation;
 use ruff_macros::{derive_message_formats, violation};
+use ruff_python_ast::token_kind::TokenKind;
 use rustpython_parser::ast::Location;
-use rustpython_parser::Tok;
 
 use super::LogicalLine;
 
@@ -253,8 +253,8 @@ pub(crate) fn indentation(
         ));
     }
     let indent_expect = prev_logical_line
-        .and_then(|prev_logical_line| prev_logical_line.tokens().trimmed().last())
-        .map_or(false, |t| t.kind() == &Tok::Colon);
+        .and_then(|prev_logical_line| prev_logical_line.tokens_trimmed().last())
+        .map_or(false, |t| t.kind() == TokenKind::Colon);
 
     if indent_expect && indent_level <= prev_indent_level.unwrap_or(0) {
         diagnostics.push((