--- a/crates/ruff/src/rules/pycodestyle/rules/logical_lines/whitespace_before_parameters.rs
+++ b/crates/ruff/src/rules/pycodestyle/rules/logical_lines/whitespace_before_parameters.rs
@@ -1,27 +1,36 @@
-use rustpython_parser::ast::Location;
-use rustpython_parser::Tok;
-
 use ruff_diagnostics::{AlwaysAutofixableViolation, Diagnostic, Edit};
 use ruff_macros::{derive_message_formats, violation};
+use ruff_python_ast::token_kind::TokenKind;
 use ruff_python_ast::types::Range;
+use rustpython_parser::ast::Location;
 
 use super::LogicalLineTokens;
 
 #[violation]
 pub struct WhitespaceBeforeParameters {
-    pub bracket: String,
+    pub bracket: TokenKind,
+}
+
+impl WhitespaceBeforeParameters {
+    fn bracket_text(&self) -> char {
+        match self.bracket {
+            TokenKind::Lpar => '(',
+            TokenKind::Lsqb => '[',
+            _ => unreachable!(),
+        }
+    }
 }
 
 impl AlwaysAutofixableViolation for WhitespaceBeforeParameters {
     #[derive_message_formats]
     fn message(&self) -> String {
-        let WhitespaceBeforeParameters { bracket } = self;
-        format!("Whitespace before {bracket}")
+        let bracket = self.bracket_text();
+        format!("Whitespace before '{bracket}'")
     }
 
     fn autofix_title(&self) -> String {
-        let WhitespaceBeforeParameters { bracket } = self;
-        format!("Removed whitespace before {bracket}")
+        let bracket = self.bracket_text();
+        format!("Removed whitespace before '{bracket}'")
     }
 }
 
@@ -33,28 +42,26 @@ pub(crate) fn whitespace_before_parameters(
     let mut diagnostics = vec![];
     let previous = tokens.first().unwrap();
 
-    let mut pre_pre_kind: Option<&Tok> = None;
+    let mut pre_pre_kind: Option<TokenKind> = None;
     let mut prev_token = previous.kind();
     let mut prev_end = previous.end();
 
     for token in tokens {
         let kind = token.kind();
 
-        if matches!(kind, Tok::Lpar | Tok::Lsqb)
-            && token.start() != prev_end
+        if matches!(kind, TokenKind::Lpar | TokenKind::Lsqb)
             && matches!(
                 prev_token,
-                Tok::Name { .. } | Tok::Rpar | Tok::Rsqb | Tok::Rbrace
+                TokenKind::Name | TokenKind::Rpar | TokenKind::Rsqb | TokenKind::Rbrace
             )
-            && (pre_pre_kind != Some(&Tok::Class))
+            && (pre_pre_kind != Some(TokenKind::Class))
+            && token.start() != prev_end
         {
             let start = Location::new(prev_end.row(), prev_end.column());
             let end = token.end();
             let end = Location::new(end.row(), end.column() - 1);
 
-            let kind: WhitespaceBeforeParameters = WhitespaceBeforeParameters {
-                bracket: kind.to_string(),
-            };
+            let kind: WhitespaceBeforeParameters = WhitespaceBeforeParameters { bracket: kind };
 
             let mut diagnostic = Diagnostic::new(kind, Range::new(start, end));