This PR introduces a new `TokenKind` enum that only represents the token's kind without the parsed data (big int literal, string...). This is useful when you want a small and copyable representation over storing a reference to `Tok`. 

Using `TokenKind` can further help with memory consumption and better cache locality because the kind is only `1byte` large compared to the 40 bytes of a `Tok`. 

This PR further changes the representation of `Tokens` from a vector of structs (`Vec<(Location, Token, Location)>`) to a struct of vectors (`kinds: Vec<TokenKind>`, `starts: Vec<Location>`, `ends: Vec<Location>`). This should reduce memory consumption because it avoids padding `TokenKind` to 4 bytes (3 wasted bytes for each token).

## Benchmark

```
group                                      no-logical                             pr3736                                 pr3745
-----                                      ----------                             ------                                 ------
linter/all-rules/large/dataset.py          1.00      8.5±0.01ms     4.8 MB/sec    1.03      8.8±0.06ms     4.6 MB/sec    1.02      8.7±0.11ms     4.7 MB/sec
linter/all-rules/numpy/ctypeslib.py        1.00      2.1±0.01ms     7.8 MB/sec    1.03      2.2±0.02ms     7.6 MB/sec    1.03      2.2±0.00ms     7.5 MB/sec
linter/all-rules/numpy/globals.py          1.00    247.9±3.25µs    11.9 MB/sec    1.06    262.9±2.80µs    11.2 MB/sec    1.09    269.7±1.99µs    10.9 MB/sec
linter/all-rules/pydantic/types.py         1.00      3.7±0.04ms     7.0 MB/sec    1.05      3.8±0.03ms     6.6 MB/sec    1.04      3.8±0.02ms     6.7 MB/sec
linter/default-rules/large/dataset.py      1.00      4.7±0.01ms     8.7 MB/sec    1.10      5.1±0.01ms     7.9 MB/sec    1.04      4.9±0.09ms     8.3 MB/sec
linter/default-rules/numpy/ctypeslib.py    1.00    999.5±5.57µs    16.7 MB/sec    1.09  1094.1±21.26µs    15.2 MB/sec    1.07  1066.8±15.69µs    15.6 MB/sec
linter/default-rules/numpy/globals.py      1.00    101.0±0.41µs    29.2 MB/sec    1.21    122.7±0.41µs    24.0 MB/sec    1.23    124.5±0.68µs    23.7 MB/sec
linter/default-rules/pydantic/types.py     1.00      2.1±0.01ms    11.9 MB/sec    1.09      2.3±0.01ms    10.9 MB/sec    1.07      2.3±0.03ms    11.2 MB/sec

```

Surprisingly, performance only changes a litle. I'm undecided if we should merge this change. The main advantage is the reduced memory consumption (for which we don't have a good measuring setup)